Moved threaded version to normal version
