+++ src/rf24.c
@@ -1,11 +1,24 @@
 #include <pthread.h>
+#include <stdio.h>
+#include <poll.h>
+#include <stdlib.h>
+#include <fcntl.h>
+#include <string.h>
+#include <unistd.h>
+#include "rf24.h"
+#include "gpio.h"
+#include "spi.h"
 #include "nRF24L01.h"
-#include "RF24_config.h"
-#include "RF24.h"
-#include "interrupts.h"
+#include "tsqueue.h"
+#include "compatibility.h"
 
 #define SPI_BITS 8
 #define SPI_MODE 0
+#define GPIO_FILE_MAXLEN  32
+#define POLL_TIMEOUT    1000
+#define RDBUF_LEN   5
+#define PACKET_BUFFER_SIZE 10
+#define ISR_PIN 24
 
 #define is_rx_fifo_empty() (read_register(FIFO_STATUS) & RX_EMPTY)
 #define is_tx_fifo_empty() (read_register(FIFO_STATUS) & TX_EMPTY)
@@ -14,6 +27,11 @@
 #define rf24_testRPD() rf24_testCarrierDetect()
 #define pipe0_is_set() (pipe0_status & 0x01)
 #define auto_ACK_occurred() (pipe0_status & 0x02)
+
+typedef struct packet {
+  uint8_t len;
+  uint8_t *payload;
+} Packet;
 
 SPIState *spi;
 uint8_t enable_pin; /**< "Chip Enable" pin, activates the RX or TX role, unused on rpi */
@@ -33,64 +51,57 @@ uint8_t pipe234_lsb[3];
 uint8_t transmit_address[5];
 uint8_t addr_width;
 pthread_t int_thread;
+pthread_mutex_t radio_lock;
+TSQueue *packets;
 /****************************************************************************/
   // Minimum ideal SPI bus speed is 2x data rate
   // If we assume 2Mbs data rate and 16Mhz clock, a
   // divider of 4 is the minimum we want.
   // CLK:BUS 8Mhz:2Mhz, 16Mhz:4Mhz, or 20Mhz:5Mhz
 
-static const char rf24_datarate_e_str_0[] PROGMEM = "1MBPS";
-static const char rf24_datarate_e_str_1[] PROGMEM = "2MBPS";
-static const char rf24_datarate_e_str_2[] PROGMEM = "250KBPS";
-static const char * const rf24_datarate_e_str_P[] PROGMEM = {
+static const char rf24_datarate_e_str_0[] = "1MBPS";
+static const char rf24_datarate_e_str_1[] = "2MBPS";
+static const char rf24_datarate_e_str_2[] = "250KBPS";
+static const char * const rf24_datarate_e_str_P[] = {
   rf24_datarate_e_str_0, 
   rf24_datarate_e_str_1, 
   rf24_datarate_e_str_2, 
 };
-static const char rf24_model_e_str_0[] PROGMEM = "nRF24L01";
-static const char rf24_model_e_str_1[] PROGMEM = "nRF24L01+";
-static const char * const rf24_model_e_str_P[] PROGMEM = {
+static const char rf24_model_e_str_0[] = "nRF24L01";
+static const char rf24_model_e_str_1[] = "nRF24L01+";
+static const char * const rf24_model_e_str_P[] = {
   rf24_model_e_str_0, 
   rf24_model_e_str_1, 
 };
-static const char rf24_crclength_e_str_0[] PROGMEM = "Disabled";
-static const char rf24_crclength_e_str_1[] PROGMEM = "8 bits";
-static const char rf24_crclength_e_str_2[] PROGMEM = "16 bits";
-static const char * const rf24_crclength_e_str_P[] PROGMEM = {
+static const char rf24_crclength_e_str_0[] = "Disabled";
+static const char rf24_crclength_e_str_1[] = "8 bits";
+static const char rf24_crclength_e_str_2[] = "16 bits";
+static const char * const rf24_crclength_e_str_P[] = {
   rf24_crclength_e_str_0, 
   rf24_crclength_e_str_1, 
   rf24_crclength_e_str_2, 
 };
-static const char rf24_pa_dbm_e_str_0[] PROGMEM = "PA_MIN";
-static const char rf24_pa_dbm_e_str_1[] PROGMEM = "PA_GPIO_LOW";
-static const char rf24_pa_dbm_e_str_2[] PROGMEM = "PA_GPIO_HIGH";
-static const char rf24_pa_dbm_e_str_3[] PROGMEM = "PA_MAX";
-static const char * const rf24_pa_dbm_e_str_P[] PROGMEM = { 
+static const char rf24_pa_dbm_e_str_0[] = "PA_MIN";
+static const char rf24_pa_dbm_e_str_1[] = "PA_GPIO_LOW";
+static const char rf24_pa_dbm_e_str_2[] = "PA_GPIO_HIGH";
+static const char rf24_pa_dbm_e_str_3[] = "PA_MAX";
+static const char * const rf24_pa_dbm_e_str_P[] = { 
   rf24_pa_dbm_e_str_0, 
   rf24_pa_dbm_e_str_1, 
   rf24_pa_dbm_e_str_2, 
   rf24_pa_dbm_e_str_3, 
 };
 
-static const uint8_t pipe_addr[] PROGMEM = {
+static const uint8_t pipe_addr[] = {
   RX_ADDR_P0, RX_ADDR_P1, RX_ADDR_P2, RX_ADDR_P3, RX_ADDR_P4, RX_ADDR_P5
 };
-static const uint8_t pipe_payload_len[] PROGMEM = {
+static const uint8_t pipe_payload_len[] = {
   RX_PW_P0, RX_PW_P1, RX_PW_P2, RX_PW_P3, RX_PW_P4, RX_PW_P5
 };
-static const uint8_t pipe_enable[] PROGMEM = {
+static const uint8_t pipe_enable[] = {
   ERX_P0, ERX_P1, ERX_P2, ERX_P3, ERX_P4, ERX_P5
 };
-
-uint8_t * reverse_address(uint8_t *address){
-  uint8_t i = 0, j = addr_width - 1, temp = 0;
-  while (i < j){
-    temp = address[i];
-    address[i++] = address[j];
-    address[j--] = temp;
-  }
-  return address;
-}
+void *radio_isr_thread();
 
 /***********************/
 /* Register functions  */
@@ -211,8 +222,32 @@ void transmit_payload(const void* buf, uint8_t len) {
   write_register(CONFIG, (read_register(CONFIG) & ~PRIM_RX)); /* Toggle RX/TX mode */
   write_payload(buf, len); /* Write the payload to the TX FIFO */
   enable_radio(); /* Pulse radio on CE pin to TX one packet from FIFO */
-  delayMicroseconds(WRITE_DELAY);
+  microSleep(WRITE_DELAY);
   disable_radio();
+}
+
+/*********************/
+/* Address functions */
+/*********************/
+uint8_t *reverse_address(uint8_t *address){
+  uint8_t i = 0, j = addr_width - 1, temp = 0;
+  while (i < j){
+    temp = address[i];
+    address[i++] = address[j];
+    address[j--] = temp;
+  }
+  return address;
+}
+
+uint8_t rf24_setAddressWidth(uint8_t address_width){
+  if (address_width > MAX_ADDR_WIDTH || address_width < MIN_ADDR_WIDTH) return 0;
+  write_register(AW, address_width);
+  addr_width = address_width;
+  return addr_width;
+}
+
+uint8_t rf24_getAddressWidth(){
+  return read_register(AW);
 }
 
 void setTXAddress(uint8_t *addr) {
@@ -234,6 +269,10 @@ void rf24_setRXAddressOnPipe(uint8_t *address, uint8_t pipe) {
   write_register(pipe_payload_len[pipe], payload_len); /* Set payload len and enable */
   write_register(EN_RXADDR, (read_register(EN_RXADDR) | pipe_enable[pipe]));
 }
+
+/***************************/
+/* TX Rate/Power functions */
+/***************************/
 
 void rf24_setDataRate(rf24_datarate_e speed) {
   uint8_t setup = read_register(RF_SETUP);
@@ -287,6 +326,10 @@ rf24_pa_dbm_e rf24_getPALevel() {
   }
 }
 
+/*****************/
+/* CRC Functions */
+/*****************/
+
 void rf24_setCRCLength(rf24_crclength_e length) {
   uint8_t config = read_register(CONFIG) & ~CRC_BITS; /* Clear CRC bits */
   switch(length){
@@ -295,10 +338,6 @@ void rf24_setCRCLength(rf24_crclength_e length) {
     case(RF24_CRC_16): config |= EN_CRC_16; break; /* Enable 16bit CRC */
   }
   write_register(CONFIG, config);
-}
-
-void rf24_disableCRC() {
-  rf24_setCRCLength(RF24_CRC_DISABLED);
 }
 
 rf24_crclength_e rf24_getCRCLength() {
@@ -316,17 +355,6 @@ bool isPVariant() {
 
 void rf24_setRetries(uint8_t delay, uint8_t count) {
  write_register(SETUP_RETR, (delay&0xf)<<ARD | (count&0xf)<<ARC);
-}
-
-uint8_t rf24_setAddressWidth(uint8_t address_width){
-  if (address_width > MAX_ADDR_WIDTH || address_width < MIN_ADDR_WIDTH) return 0;
-  write_register(AW, address_width);
-  addr_width = address_width;
-  return addr_width;
-}
-
-uint8_t rf24_getAddressWidth(){
-  return read_register(AW);
 }
 
 void rf24_setChannel(uint8_t channel) {
@@ -361,7 +389,7 @@ uint8_t rf24_init_radio(char *spi_device, uint32_t spi_speed, uint8_t cepin) {
   // Enabling 16b CRC is by far the most obvious case if the wrong timing is used - or skipped.
   // Technically we require 4.5ms + 14us as a worst case. We'll just call it 5ms for good measure.
   // WARNING: Delay is based on P-variant whereby non-P *may* require different timing.
-  delay(5);
+  milliSleep(5);
   // Set 1500uS (minimum for 32B payload in ESB@250KBPS) timeouts, to make testing a little easier
   // WARNING: If this is ever lowered, either 250KBS mode with AA is broken or maximum packet
   // sizes must never be used. See documentation for a more complete explanation.
@@ -396,10 +424,14 @@ uint8_t rf24_init_radio(char *spi_device, uint32_t spi_speed, uint8_t cepin) {
   // spectrum.
   rf24_setChannel(76);
 
+  /* Set default address width to 5bytes */
+  rf24_setAddressWidth(5);
+
   // Flush buffers
   flush_rx();
   flush_tx();
-  pthread_create(&int_thread, NULL, interrupt_wait, (void *)24);
+  pthread_create(&int_thread, NULL, radio_isr_thread, NULL);
+  packets = tsq_create(PACKET_BUFFER_SIZE);
   return 1;
 }
 
@@ -414,7 +446,7 @@ void rf24_startListening() {
   if (PIPE0_SET && PIPE0_AUTO_ACKED) 
     write_register_bytes(RX_ADDR_P0, reverse_address(pipe0_address), addr_width);
   enable_radio();
-  delayMicroseconds(LISTEN_DELAY); /* wait for the radio to come up */
+  microSleep(LISTEN_DELAY); /* wait for the radio to come up */
 }
 
 void rf24_stopListening() {
@@ -425,12 +457,12 @@ void rf24_stopListening() {
 
 void rf24_powerDown() {
   write_register(CONFIG, (read_register(CONFIG) & ~PWR_UP));
-  delayMicroseconds(POWER_DOWN_DELAY);
+  microSleep(POWER_DOWN_DELAY);
 }
 
 void rf24_powerUp() {
   write_register(CONFIG, (read_register(CONFIG) | PWR_UP));
-  delayMicroseconds(POWER_UP_DELAY);
+  microSleep(POWER_UP_DELAY);
 }
 
 bool rf24_available(uint8_t* pipe_num) {
@@ -443,16 +475,14 @@ bool rf24_available(uint8_t* pipe_num) {
   return result;
 }
 
-uint8_t rf24_recv(void* buf, uint8_t len, uint8_t flags) {
-  uint8_t payload_len = (dyn_payloads_set ? get_dyn_payload_len() : MAX_PAYLOAD_LEN);
-  if (payload_len > MAX_PAYLOAD_LEN){
-    flush_rx(); /* Invalid payload needs flushing */
-    return -1;
-  }
-  read_payload(buf, len, payload_len); /* Fetch the payload */
-  if (is_rx_fifo_empty()) /* Clear status bit if there are no more payloads */
-    write_register(STATUS, RX_DR);
-  return payload_len;
+uint8_t rf24_recv(void* buf, uint8_t len, uint8_t block) {
+  Packet * p = tsq_remove(packets, block);
+  if (p == NULL) return 0; /* No packet available (nonblocking) */
+  memcpy(buf, p->payload, (p->len > len ? len : p->len));
+  uint8_t p_len = p->len; /* Save len whilst we free the memory */
+  free(p->payload);
+  free(p);
+  return p_len;
 }
 
 bool rf24_send(uint8_t *addr, const void* buf, uint8_t len) {
@@ -471,14 +501,14 @@ bool rf24_write(const void* buf, uint8_t len) {
 
   uint8_t observe_tx;
   uint8_t status;
-  uint32_t sent_at = __millis();
+  uint32_t sent_at = millis();
   const uint32_t timeout = 500; //ms to wait for timeout
   do
   {
     status = read_register_bytes(OBSERVE_TX, &observe_tx, 1);
-    IF_SERIAL_DEBUG(printf("%x", observe_tx));
+    DEBUG_PRINT(printf("%x", observe_tx));
   }
-  while(! (status & (TX_DS | MAX_RT)) && (__millis() - sent_at < timeout));
+  while(! (status & (TX_DS | MAX_RT)) && (millis() - sent_at < timeout));
 
   bool tx_ok, tx_fail;
   rf24_getStatus(&tx_ok, &tx_fail, &ack_payload_available);
@@ -486,13 +516,13 @@ bool rf24_write(const void* buf, uint8_t len) {
   //printf("%u%u%u\r\n", tx_ok, tx_fail, ack_payload_available);
 
   result = tx_ok;
-  IF_SERIAL_DEBUG(printf("%s\n", result ? "...OK." : "...Failed"));
+  DEBUG_PRINT(printf("%s\n", result ? "...OK." : "...Failed"));
 
   // Handle the ack packet
   if (ack_payload_available) {
     ack_payload_length = get_dyn_payload_len();
-    IF_SERIAL_DEBUG(printf("[AckPacket]/"));
-    IF_SERIAL_DEBUG(printf("%i\n", ack_payload_length));
+    DEBUG_PRINT(printf("[AckPacket]/"));
+    DEBUG_PRINT(printf("%i\n", ack_payload_length));
   }
   return result;
 }
@@ -505,6 +535,12 @@ void rf24_getStatus(bool *tx_ok, bool *tx_fail, bool *rx_ready) {
   if (rx_ready) *rx_ready = status & RX_DR;
 }
 
+void rf24_peekStatus(bool *tx_ok, bool *tx_fail, bool *rx_ready) {
+  uint8_t status = check_status();
+  if (tx_ok) *tx_ok = status & TX_DS;
+  if (tx_fail) *tx_fail = status & MAX_RT;
+  if (rx_ready) *rx_ready = status & RX_DR;
+}
 void rf24_autoACKPacket(){
     write_register_bytes(RX_ADDR_P0, reverse_address(transmit_address), addr_width);
     write_register(RX_PW_P0, (payload_len < MAX_PAYLOAD_LEN ? payload_len : MAX_PAYLOAD_LEN));
@@ -538,7 +574,7 @@ void rf24_enableAckPayload() {
       write_register(FEATURE, (EN_ACK_PAY | EN_DPL));
     }
   }
-  IF_SERIAL_DEBUG(printf("FEATURE=%i\r\n", read_register(FEATURE)));
+  DEBUG_PRINT(printf("FEATURE=%i\r\n", read_register(FEATURE)));
   /* Enable dynamic payload on pipes 0 */
   write_register(DYNPD, (read_register(DYNPD) | DPL_P0));
 }
@@ -617,10 +653,10 @@ void rf24_printDetails() {
   printf("SPI device\t = %s\r\n", spidevice);
   printf("SPI speed\t = %d\r\n", spispeed);
   printf("CE GPIO\t = %d\r\n", enable_pin);
-  printf("Data Rate\t = %s\r\n", pgm_read_word(&rf24_datarate_e_str_P[rf24_getDataRate()]));
-  printf("Model\t\t = %s\r\n", pgm_read_word(&rf24_model_e_str_P[isPVariant()]));
-  printf("CRC Length\t = %s\r\n", pgm_read_word(&rf24_crclength_e_str_P[rf24_getCRCLength()]));
-  printf("PA Power\t = %s\r\n", pgm_read_word(&rf24_pa_dbm_e_str_P[rf24_getPALevel()]));
+  printf("Data Rate\t = %s\r\n", rf24_datarate_e_str_P[rf24_getDataRate()]);
+  printf("Model\t\t = %s\r\n", rf24_model_e_str_P[isPVariant()]);
+  printf("CRC Length\t = %s\r\n", rf24_crclength_e_str_P[rf24_getCRCLength()]);
+  printf("PA Power\t = %s\r\n", rf24_pa_dbm_e_str_P[rf24_getPALevel()]);
   print_status(check_status());
   print_address_register("RX_ADDR_P0-1", RX_ADDR_P0, 2);
   print_byte_register("RX_ADDR_P2-5", RX_ADDR_P2);
@@ -632,4 +668,75 @@ void rf24_printDetails() {
   print_byte_register("RF_SETUP", RF_SETUP);
   print_byte_register("CONFIG", CONFIG);
   print_byte_register("DYNPD/FEATURE", DYNPD);
+}
+
+int setup_isr_thread(int pin) {
+  char gpio_file[GPIO_FILE_MAXLEN];
+  gpio_open(pin, GPIO_IN);
+  gpio_enable_edge(pin, GPIO_FALLING_EDGE);
+  memset(gpio_file, 0x00, GPIO_FILE_MAXLEN);
+  snprintf(gpio_file, GPIO_FILE_MAXLEN - 1, "/sys/class/gpio/gpio%d/value", pin);
+  return open(gpio_file, O_RDONLY);
+}
+
+void retrieve_packets(){
+  uint8_t payload_len;
+  Packet *packet;
+  while (!is_rx_fifo_empty()){
+    payload_len = (dyn_payloads_set ? get_dyn_payload_len() : MAX_PAYLOAD_LEN);
+    if (payload_len > MAX_PAYLOAD_LEN){
+      flush_rx(); /* Invalid payload needs flushing */
+      continue;
+    }
+    packet = (Packet*)malloc(sizeof(Packet));
+    if (packet == NULL) return;
+    packet->payload = (uint8_t *)malloc(payload_len);
+    if (packet->payload == NULL) {
+      free(packet);
+      return;
+    }
+    packet->len = payload_len;
+    read_payload(packet->payload, payload_len, payload_len); /* Fetch the payload */
+    tsq_add(packets, packet, 0); /* Don't block, if the q is full it's dropped */
+  }
+  /* Clear status bit if there are no more payloads */
+  write_register(STATUS, RX_DR);
+}
+
+void process_radio_interrupt() {
+  bool tx_ok, tx_fail, pkt_avail;
+  rf24_peekStatus(&tx_ok, &tx_fail, &pkt_avail);
+  if (pkt_avail) retrieve_packets();
+}
+
+void *radio_isr_thread() {
+  int fd, result;
+  struct pollfd pfd;
+  char rdbuf[RDBUF_LEN];
+  fd = setup_isr_thread(ISR_PIN);
+  if (fd < 0) {
+    perror("gpio_file");
+    return (void *)-1;
+  }
+  pfd.fd = fd;
+  pfd.events = POLLPRI;
+
+  while(1) {
+    memset(rdbuf, 0x00, RDBUF_LEN);
+    lseek(fd, 0, SEEK_SET);
+    result = poll(&pfd, 1, -1);
+    if (result < 0) {
+      perror("poll()");
+      close(fd);
+      return (void *)3;
+    }
+    result = read(fd, rdbuf, RDBUF_LEN);
+    if (result < 0) {
+      perror("read()");
+      return (void *)4;
+    }
+    process_radio_interrupt();
+  }
+  close(fd);
+  return (void *)0;
 }
